flatten Auth facade using zod() Schema→Zod interop
